JiHeeChoi22.github.io/Automation/selenium_test_project/requirements.txt
selenium==4.0.0
pytest==7.0.0
pytest-html==3.1.1
python-dotenv==0.19.0
webdriver-manager==3.5.2
pytest-xdist==2.5.0 